% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_import_functions.R
\name{import_bis_credit_to_gdp_ratios}
\alias{import_bis_credit_to_gdp_ratios}
\title{This function imports credit to GDP ratios from BIS format}
\usage{
import_bis_credit_to_gdp_ratios(
  file_path,
  my_frequency = "Quarterly",
  my_borrowing_sector = "Private non-financial sector",
  my_lending_sector = "All sectors"
)
}
\arguments{
\item{file_path}{string a file path to the source data file}

\item{my_frequency}{time frequency of the data (default is Quarterly),
available options are:
\itemize{
 \item Quarterly
}}

\item{my_borrowing_sector}{(default is Private non-financial sector),
available options are:
\itemize{
 \item Private non-financial sector
}}

\item{my_lending_sector}{(default is All sectors),
available options are:
\itemize{
 \item All sectors
}}
}
\description{
This function imports credit to GDP ratios from BIS format
}
