% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_import_functions.R
\name{import_bis_debt_sec}
\alias{import_bis_debt_sec}
\title{This function imports debt securities from BIS format}
\usage{
import_bis_debt_sec(
  file_path,
  my_frequency = NULL,
  my_measure = NULL,
  my_country = NULL,
  my_issuer_sector_immediate_borrower = NULL,
  my_issuer_sector_ultimate_borrower = NULL,
  my_issue_market = NULL,
  my_issue_currency_group = NULL,
  my_issue_currency = NULL,
  my_original_maturity = NULL,
  my_remaining_maturity = NULL,
  my_rate_type = NULL,
  my_default_risk = NULL,
  my_collateral_type = NULL
)
}
\arguments{
\item{file_path}{string a file path to the source data file}

\item{my_frequency}{time frequency of the data (default is NULL),
available options are:
\itemize{
 \item Quarterly
}}

\item{my_measure}{(default is NULL),
available options are:
\itemize{
 \item Amounts outstanding
 \item Net issues
 \item Gross issues
}}

\item{my_issuer_sector_immediate_borrower}{(default is NULL),
available options are:
\itemize{
 \item All issuers
 \item Financial corporations
 \item Private banks
 \item International institutions
 \item Public banks
 \item General government
 \item Non-financial corporations
 \item Private other financial institutions
 \item Central bank
 \item Public other financial institutions
}}

\item{my_issuer_sector_ultimate_borrower}{(default is NULL),
available options are:
\itemize{
 \item All issuers
 \item Financial corporations
 \item Private banks
 \item International institutions
 \item Public banks
 \item General government
 \item Non-financial corporations
 \item Private other financial institutions
 \item Central bank
 \item Public other financial institutions
}}

\item{my_issue_market}{(default is NULL),
available options are:
\itemize{
 \item All markets
 \item Domestic market
 \item International markets
}}

\item{my_issue_currency_group}{(default is NULL),
available options are:
\itemize{
 \item All currencies
 \item Foreign currencies
 \item Domestic currency
}}

\item{my_issue_currency}{(default is NULL),
There are 65 available options, the special are:
\itemize{
 \item Total all currencies
 \item Sum of ECU, Euro and legacy currencies now included in the Euro
}}
}
\description{
This function imports debt securities from BIS format
}
