% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_import_functions.R
\name{import_bis_lbs}
\alias{import_bis_lbs}
\title{This function imports locational banking statistics from BIS format}
\usage{
import_bis_lbs(
  file_path,
  my_frequency = NULL,
  my_measure = NULL,
  my_balance_sheet_position = NULL,
  my_type_of_instruments = NULL,
  my_currency_denomination = NULL,
  my_currency_type_of_reporting_country = NULL,
  my_parent_country = NULL,
  my_type_of_reporting_institutions = NULL,
  my_reporting_country = NULL,
  my_counterparty_sector = NULL,
  my_counterparty_country = NULL,
  my_position_type = NULL,
  pivot_to_long = FALSE
)
}
\arguments{
\item{file_path}{string a file path to the source data file}

\item{my_frequency}{time frequency of the data (default is NULL),
available options are:
\itemize{
 \item Quarterly
}}

\item{my_measure}{(default is NULL),
available options are:
\itemize{
 \item FX and break adjusted change (BIS calculated)
 \item Amounts outstanding / Stocks
}}

\item{my_balance_sheet_position}{(default is NULL)
available options are:
\itemize{
 \item Total claims
 \item Total liabilities
}}

\item{my_type_of_instruments}{(default is NULL)
available options are:
\itemize{
 \item All instruments
 \item Debt securities
 \item Loans and deposits
 \item Other instruments
 \item Unallocated by instrument
 \item Debt securities, short-term
}}

\item{my_currency_denomination}{(default is NULL),
available options are:
\itemize{
 \item Swiss Franc
 \item Euro
 \item Pound Sterling
 \item Japanese Yen
 \item US Dollar
 \item All currencies
 \item All currencies excluding USD, EUR, JPY, CHF and GBP
 \item Unallocated currencies
}}

\item{my_currency_type_of_reporting_country}{(default is NULL),
available options are:
\itemize{
 \item All currencies (=D+F+U)
 \item Foreign currency (ie currencies foreign to bank location country)
 \item Domestic currency (ie currency of bank location country)
 \item Unclassified currency
}}

\item{my_parent_country}{string}

\item{my_type_of_reporting_institutions}{(default is NULL),
available options are:
\itemize{
 \item All reporting banks/institutions (domestic, foreign, consortium and unclassified)
 \item Foreign branches
 \item Domestic banks
 \item Foreign subsidiaries
}}

\item{my_reporting_country}{string}

\item{my_counterparty_sector}{(default is NULL),
available options are:
\itemize{
 \item All sectors
 \item Banks, total
 \item Non-financial corporations
 \item Non-bank financial institutions
 \item General government
 \item Households and NPISHs
 \item Banks, related offices
 \item Banks, central banks
 \item Non-banks, total
 \item Non-financial sectors
 \item Unallocated by sector

}}

\item{my_position_type}{(default is NULL),
available options are:
\itemize{
 \item All
 \item Cross-border
 \item Local
 \item Unallocated
}}

\item{pivot_to_long}{reshape the data to long format? Default is FALSE}

\item{counterparty_country}{string}
}
\description{
This function imports locational banking statistics from BIS format
}
