% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_import_functions.R
\name{import_bis_policy_rates}
\alias{import_bis_policy_rates}
\title{This function imports policy rates from BIS format}
\usage{
import_bis_policy_rates(file_path)
}
\arguments{
\item{file_path}{string a file path to the source data file}
}
\description{
This function imports policy rates from BIS format
}
