% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_import_functions.R
\name{import_bis_total_credit}
\alias{import_bis_total_credit}
\title{This function imports total credit to private sector from BIS format}
\usage{
import_bis_total_credit(
  file_path = NULL,
  my_frequency = NULL,
  my_borrowing_sector = NULL,
  my_lending_sector = NULL,
  my_valuation = NULL,
  my_unit_type = NULL,
  my_type_of_adjustment = NULL,
  pivot_to_long = FALSE
)
}
\arguments{
\item{file_path}{string a file path to the source data file}

\item{my_frequency}{time frequency of the data (default is NULL),
available options are:
\itemize{
 \item Quarterly
}}

\item{my_borrowing_sector}{(default is NULL),
available options are:
\itemize{
 \item Non financial sector
 \item General government
 \item Households & NPISHs
 \item Non-financial corporations
 \item Private non-financial sector
}}

\item{my_lending_sector}{(default is NULL),
available options are:
\itemize{
 \item All sectors
 \item Banks, domestic
}}

\item{my_valuation}{(default is NULL), toggles the
difference between credit and debt. available options are:
\itemize{
 \item Nominal value
 \item Market value
}}

\item{my_unit_type}{(default is NULL),
available options are:
\itemize{
 \item US Dollar
 \item Percentage of GDP
 \item Percentage of GDP-PPP
 \item Domestic currency
}}

\item{my_type_of_adjustment}{(default is NULL),
available options are:
\itemize{
 \item Adjusted for breaks
 \item Unadjusted for breaks
}}

\item{pivot_to_long}{reshape the data to long format? Default is FALSE}
}
\description{
This function imports total credit to private sector from BIS format
}
