% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boi_import_functions.R
\name{import_boi_oracle_finrep_data}
\alias{import_boi_oracle_finrep_data}
\title{Import Oracle format financial report data}
\usage{
import_boi_oracle_finrep_data(filepath = NULL, data_frequency)
}
\arguments{
\item{filepath}{the path to financial report data (in rds format)}

\item{data_frequency}{filtering parameter. can be one of:
\itemize{
\item {annual}
\item {semiannual}
\item {quarterly}
}
The default}
}
\description{
This function imports financial report data from Oracle format
}
